+++ lib/apigateway/lib/model/create-gateway-details.ts
@@ -87,6 +87,17 @@ Example: {@code {\"Operations\": {\"CostCenter\": \"42\"}}}
    * An array of CA bundles that should be used on the Gateway for TLS validation.
    */
   "caBundles"?: Array<model.CaBundle>;
+  /**
+   * Determines whether the gateway has an IPv4 or IPv6 address assigned to it, or both.
+   * {@code IPV4} means the gateway will only have an IPv4 address assigned to it, and {@code IPV6} means the gateway will
+   * only have an {@code IPv6} address assigned to it. {@code DUAL_STACK} means the gateway will have both an IPv4 and IPv6
+   * address assigned to it.
+   * Example: {@code IPV4} or {@code IPV6} or {@code DUAL_STACK}
+   *
+   */
+  "ipMode"?: string;
+  "ipv6AddressConfiguration"?: model.Ipv6AddressConfiguration;
+  "ipv4AddressConfiguration"?: model.Ipv4AddressConfiguration;
 }
 
 export namespace CreateGatewayDetails {
@@ -107,6 +118,13 @@ export namespace CreateGatewayDetails {
           ? obj.caBundles.map(item => {
               return model.CaBundle.getJsonObj(item);
             })
+          : undefined,
+
+        "ipv6AddressConfiguration": obj.ipv6AddressConfiguration
+          ? model.Ipv6AddressConfiguration.getJsonObj(obj.ipv6AddressConfiguration)
+          : undefined,
+        "ipv4AddressConfiguration": obj.ipv4AddressConfiguration
+          ? model.Ipv4AddressConfiguration.getJsonObj(obj.ipv4AddressConfiguration)
           : undefined
       }
     };
@@ -130,6 +148,13 @@ export namespace CreateGatewayDetails {
           ? obj.caBundles.map(item => {
               return model.CaBundle.getDeserializedJsonObj(item);
             })
+          : undefined,
+
+        "ipv6AddressConfiguration": obj.ipv6AddressConfiguration
+          ? model.Ipv6AddressConfiguration.getDeserializedJsonObj(obj.ipv6AddressConfiguration)
+          : undefined,
+        "ipv4AddressConfiguration": obj.ipv4AddressConfiguration
+          ? model.Ipv4AddressConfiguration.getDeserializedJsonObj(obj.ipv4AddressConfiguration)
           : undefined
       }
     };