+++ lib/core/lib/model/create-vcn-details.ts
@@ -124,7 +124,7 @@ Example: {@code true}
     */
   "isIpv6Enabled"?: boolean;
   /**
-   * Indicates whether Zpr Only Mode is enforced.
+   * Indicates whether ZPR Only mode is enforced.
    *
    */
   "isZprOnly"?: boolean;