+++ lib/database/lib/model/identity-connector-details.ts
@@ -33,6 +33,7 @@ export namespace IdentityConnectorDetails {
   export enum CloudProvider {
     Azure = "AZURE",
     Gcp = "GCP",
+    Aws = "AWS",
     /**
      * This value is used if a service returns a value for this enum that is not recognized by this
      * version of the SDK.